Separate floating ui code from the rest of the implementation
